Update macOS config for Privacy Pro survey changes (#101)
Task: https://app.asana.com/0/1199333091098016/1208279675094863/f

This PR updates the macOS Privacy Pro surveys to add a hardcoded build type parameter to the survey URLs. This is a temporary measure until the survey builder supports the build type parameter.

See the linked Asana task for an easier-to-read mapping of rules to messages.