re-added workaround to not display preview for installs with no AtbInitDate (#141)
daysSinceInstalled matching attribute has a bug where the install date
is set as 01/01/0001. This means the matching attribute thinks the
browser has been installed for over 700,000 days, causing min
comparisons to incorrectly return true.

Added a "max" comparison to not shown the message under these
circumstances.

See also -
https://github.com/duckduckgo/remote-messaging-config/pull/136

This was reverted as we thought this was causing stable to break on non
UTC timezones. This was a false alarm - this message is fine. (checked
on timezones + and - from UTC)