Added windows q4 survey (for 11th Nov) (#261)
https://app.asana.com/1/137249556945/project/1207619243206445/task/1211858291694027

This PR adds the windows Q4 survey.

This config is a little janky...

- There was an ad block experiment running from 15th July - 15th Sep
- This experiment had a survey as part of it which was shown/handled
natively - the RMF framework has no way of knowing if the user has
interacted with it.
- So we decided to not show the survey to any users who installed in
that date range
- But there is no matching attribute to match against an install date
range
- The closest is matching on install atb - the atb range for the date
range above maps to `v490-7` -> `v499-6`
- But the install atb matching attribute only lets you match against a
single value
- So we need 63 individual matching attributes that we exclude the
message against with one atb value in each.

This does work, but I'm certainly going to raise an improvement
suggestion for the matching attribute.