Android: Remove Black Friday campaign (#286)
**Asana**: 

https://app.asana.com/1/137249556945/project/1207619243206445/task/1211632234788074?focus=true

**Description**:
Removes Black Friday message included in
https://github.com/duckduckgo/remote-messaging-config/pull/271

**Testing**:
Check that the survey is correctly removed based on the diff of the
original PR linked above.

NOTE: This is stack **[1]** in a stacked PR.

Base
[main](https://github.com/duckduckgo/remote-messaging-config/tree/main)
Stack 0 -
[noelia/add-black-friday-campaign](https://github.com/duckduckgo/remote-messaging-config/tree/feature/noelia/add-black-friday-campaign)
Stack 1 -
[noelia/remove-black-friday-promo](https://github.com/duckduckgo/remote-messaging-config/tree/feature/noelia/remove-black-friday-promo)
<- You are here

We are however merging directly to main. The intent is that we're
preparing a series of messages in advance, all of which should overwrite
the previous ones. This PR in particular is prepared to remove the
previously added message - see:

- https://github.com/duckduckgo/remote-messaging-config/pull/271